Replace connection/listen_args where possible
Previously we used the security object to create connection_args and listen_args
This was some unnecessary redundant state and makes it a bit harder to
change things in the future
(see https://github.com/dask/distributed/issues/3339)

Now we remove excess calls to connect and instead try to centralize
everything to the `ConnectionPool`.  Hopefully we can isolate changes to
that in the future.