Skip test_open_close_many_workers on Python 3.6
This test has caused intermittent failures in the past.
Previously we had marked it as xfail, but it would still cause CI to
break because it would cause things to hang.

In #3419 it was observed that the failure seems to only occur on Python
3.6.  This commit changes the universal xfail to a skipif for Python 3.6
and below.

We still don't know what causes the failure, other than that GC seems to
take up all of the CPU time

Fixes #3419