Cast CUDA array data to `uint8` before serializing
As we will ultimately read any data in as `uint8`, go ahead and cast to
`uint8` as part of serialization. This is a good first order check to
make sure that we are able to serialize the dat

To simplify handling of the data a bit, go ahead and cast it to `uint8`
before serializing. This makes contiguity checks trivial.