Register Dask cuDF serializers
As cuDF is gaining support for serializing using the Dask protocol as
well, make sure to register it's serializers there as well. cuDF
versions lacking this support will continue to behave the same (falling
back to pickle). However cuDF versions with this support will bypass
pickle. So only pay the cost of a host-to-device transfer.