Remove --verbose flag from CI runs
Historically we included the verbose flag because if things hung it was
useful to see which test in particular was causing the hang.  This
hasn't been so important recently.

It's somewhat annoying to scroll through all of the tests one by one.
This seems more important today than the hung test case.