Pin `numpydoc` to avoid double escaped *
Similar to dask/dask#5961, recent changes to `numpydoc` lead to function
signatures displayed using `autosummary` to have doubly-escaped `*`s.

This pins `numpydoc` to version 0.8.0 to avoid the regression until a
patch is merged upstream.

Also small changes to avoid a sphinx timeout looking for an intersphinx
inventory, and a fixed a misformatted code-block