Special case scipy.sparse's `dok_matrix`
As `dok_matrix` is a subclass of `spmatrix` and `dict`, this confuses
Dask's `register_generic` machinery. Not to mention this doesn't
actually contain any NumPy `ndarray`s. Instead it just stores
coordinates as `tuple`s in keys and data in values.

Ideally we would just pack the dictionary into the header and move on.
However as the data included is not MessagePack serializable, this
strategy does not work in practice. So simply convert the `dok_matrix`
to a `coo_matrix`, which has very similar layout and is easier for us to
serialize. When deserializing, just extract the `coo_matrix` and convert
it back to a `dok_matrix`.

This let's us bypass the oddities of the `dok_matrix` while still having
reasonably efficient serialization.