Special case `MatDescriptor` serialization
It turns out `MatDescriptor` contains some state that (while it looks
serializable) is not. In particular it is carrying around a Python
`int`, which in reality is a pointer to some resource allocated by
cuSPARSE. Attempting to reuse that pointer outside of where it is valid
causes a segfault. We fix this by simply creating a new `MatDescriptor`
object in the process where it was serialized too. Additionally we
special case pickle (until it is fixed upstream) to ensure this behavior
is avoided.

ref: https://docs.nvidia.com/cuda/cusparse/index.html#cusparseCreateMatDescr