Ensure CUDA context on `DeviceNDArray` cleanup
As the CUDA context does not seem to always be established when cleaning
up Numba `DeviceNDArray`s, make sure to get the context right before
cleanup to ensure the context is available.

This is done when allocating new CUDA frames to when using UCX. Make
sure this is handled both for Numba and old versions of RMM that also
use Numba under-the-hood. Newer versions of RMM don't need this.

Also do this when performing a host-to-device transfer where a Numba
`DeviceNDArray` is the first object created on device.