Drop `PatchedCudaArrayInterface` and usage thereof
This was used to ensure that Numba had acquired a CUDA context before
finalizing any Numba `DeviceNDArray` objects that might be used to back
a CuPy `ndarray`. As we now make sure of this when creating Numba
`DeviceNDArray` by using `weakref.finalize`, this should be correctly
handled for all CUDA objects backed by Numba `DeviceNDArray`s (not just
CuPy `ndarray`s). So there should be no need to do this for CuPy
`ndarray`s separately. Should simplify the CuPy serialization and
cleanup paths a bit.