Contain `check_deserialize` import
Sometimes this doesn't work on some machines. We are still working out
how to get this to work more reliably. For now just skip the test if
this `import` doesn't work for any reason. Also contain the `import` to
this test as other tests don't need it and we don't want the tests to
fail to run entirely.