Allow `test_ucx_deserialize` to error
Instead of trying to catch and skip certain errors, go ahead and just add a
comment to this test about how it can error on some systems. Since this no
longer blocks the full test suite from running, it is less of an issue. Plus
seeing it may give us motivation to fix either the system or the test in the
future ;)