Create futures at the beginning of scatter rather than the end
Otherwise there is a period where the user has requested that data
exists, but we haven't incremented the reference counter yet.
This way we've stated that yes, these keys are important to us, even
while we wait for the data to transfer.  This stops the client from
sending an erroneous release message to the scheduler.