Collect buffers internally first
Before calling the user provided buffer callback, collect buffers in an
internal list. That way if the mechanism of pickling needs to be
changed, the internal list can be purged before handing these to the
user. At the end of pickling, make sure the user's buffer callback is
called on each buffer in order.