Allow to pass in task keys in the restrictions
The `workers` parameter of Client.compute and Client.persist
allows to restrict the workers where the calculations of the tasks
should take place.
Advanced usages of this keyword might need passing the task key
(string) instead of the dask collection or dask task.
Internally, collections and tasks are turned into string keys
anyways.

This allows now to do things like:

	task = ... long dependencies ...
        all_dependencies = task.dask.dependencies.keys()

        workers = {d: worker_address for d in all_dependencies}
        result = client.compute(last_task, workers=workers)

to have a task and all of its dependencies calculated on a single
worker.