Merge frames in `deserialize_bytes`
It appears that we are splitting frames in `serialize_byteslist` so that
we can compress them. However we are not merging them back together
afterwards during deserialization. This can cause an exception to be
raised by serializers that expected their frames to be structured in a
particular way.

To fix this, we make sure to call `merge_frames` after `decompress` in
`deserialize_bytes`. Further we make sure to pack the `lengths` of the
original `frames` in the `header` (if not already present) in
`serialize_byteslist`. This should ensure deserializers get the original
frame structuring back when they operate on them.