Drop old frame splitting in NumPy serialization
This is leftover from when Blosc compression was still being handled
with the NumPy serialization code path itself. These days frame
splitting already happens as part of serialization more generally before
compressing data. Also this cutoff is not relevant for when splitting
occurs as that is determined by a limit in the size of buffers Blosc can
handle, which `frame_split_size` already knows. So go ahead and drop
this special casing in NumPy serialization and rely on these commonly
shared code paths that all serialization passes through.