Track `memoryview`'s `format` and `shape`
When serializing a `memoryview`, track its format and shape. Make sure
to reinstantiate the `memoryview` on the other end with that format and
shape.

It's worth noting that `memoryview`s don't actually allow `strides` to
be passed as well. So we skip this. As a result if a F-order
`memoryview` is needed, one will need to perform a copy to fix this.