Raise if a `memoryview` with `object`s is given
Trying to turn a `memoryview` of `object`s into `bytes` doesn't work
particularly well as it will contain only the memory pointers of the
Python objects without their actual data. So they won't be serialized
correctly when going over to the receiver. There isn't a good
alternative solution in this case. So just raise.