Ensure all frames are writable before returning
As users expect objects that are deserialized like NumPy arrays and
Pandas objects are mutable, make sure the frames used to build them up
are also mutable. Do this by checking if they are read-only (as is the
case for `bytes` as an example). If they are, copy them into a
`bytearray` backed `memoryview`. Otherwise pass them through as-is.