Protect profile against very large stacks
This can come up in tracebacks when exporting profiles.

This isn't quite right.  We're chopping off the calling functions of the
call stack rather than the called functions, but it's easiest to do it
this way given how we've arranged our profile processing function.  So
the top of the call stack will appear somewhat disassociated, but will
still be around.

Mostly I don't care too much about pathological cases like these, as
long as they don't ruin communication of the rest of the system.