CLN : Update super usage
On Python 3, super can be called with no arguments but on Python 2,
super needs to be called with the class name and an instance.

Because distributed does not support Python 2 on master, super usage can
be updated.

Note that this was an automated regex-based search and replace. After
the automated replace, each change was individually checked and
committed.

	modified:   distributed/client.py
	modified:   distributed/deploy/adaptive.py
	modified:   distributed/deploy/local.py
	modified:   distributed/diagnostics/progressbar.py
	modified:   distributed/http/scheduler/prometheus/__init__.py
	modified:   distributed/http/worker/prometheus.py
	modified:   distributed/nanny.py
	modified:   distributed/scheduler.py
	modified:   distributed/tests/test_scheduler.py
	modified:   distributed/threadpoolexecutor.py
	modified:   distributed/utils.py
	modified:   distributed/worker.py