Make test compatible LoopRunner
LoopRunner is designed to stop loops it started.
`loop` fixture doesn't start its IOLoop but always stops it.
This leads to conflict in the test: both LoopRunner and `loop` fixture
stop the same loop. So we need to use `loop_in_thread` fixture
since it starts and stops its IOLoop.