Configurable timeouts for worker_client and get_client.
Issue: https://github.com/dask/distributed/issues/4114

Previously, connection timeout for worker_client and get_client was
hard-coded to 3s by default.
With this change, the default timeout value is fetched from the
dask config 'distributed.comm.timeouts.connect' if set. And if
this config is not set, '3s' is taken as default.