Handle exceptions in BatchedSend
This changes exception handling in BatchedSend.

The problem: BatchedSend.send is a regular function. What happens if we
add a message to be sent, but that send fails at a later time? In the
current structure, that exception is logged and then lost to the void.

This PR makes things a bit more robust in two ways:

1. For unexpected errors, we retry a few times. In
   https://github.com/pangeo-data/pangeo/issues/788 manual retries did
   fix the issue, so we try that here automatically.
2. For CommClosedErrors or repeated unexpected errors, we close the
   BatchedSend object. By closing the BatchedSend after an exception,
   *subsequent* `BatchedSend.send` calls will fail.

We still face the issue with a potential exception happening in the
background that *aren't* followed by another `BatchedSend.send` failing.
But I think that's unavoidable given the current design.