Replace `self.executing` with `self.executing_count`
This is a redundant bit of state tracking and overcomplicates things.
Instead, similar to `Worker.tasks` this maintains a count of the
executing tasks for reporting but relies on `TaskState` objects to keep
track of the task state.

Note that the removal of the `finally:` block, specifically, is because
_if_ the key has already been released, then `self.executing_count` is
already decremented so that particular check doesn't need to be run.