Don't assume "waiting" for new dependency tasks
This fixes the sporadic error that was showing up in
test_failed_workers.py::test_worker_who_has_clears_after_failed_connection

The tests would occasionally hit a validation error when trying to
transition from waiting -> executing because the task key already
existed in `self.data`.

I thought this was because of an extra call to `release_key` somewhere,
and it _is_, but that call is _in the test_ -- it's called explicitly as
part of that test.

So when it's called but the given task still has dependent tasks, we
remove the TaskState object but leave the data in `self.data` (because
the dependents probably need it) -- but then when the task is recreated
as part of the call to `add_task` (on behalf of that depedent task), the
state is set to "waiting", because this is almost always the correct
state.

But not always, as in this case. If the data already exists on the
worker because something called `release_key` and then the task is
recreated, the state should be `memory` because the output is already in `self.data`