Create dependency TaskState as needed in gather_dep
Fixes #4231

As raised in #4231 there are a few cases where we try to gather
dependencies that don't have a corresponding TaskState object -- I think
this is probably because those objects have been recently released (or
it may be due to work-stealing, still trying to work that out
definitively).

This prevents the worker from bailing out -- and in either of the cases
above, we'll still need to ask the scheduler for more information, so we
don't need to add this particular TaskState object to the tasks dict.