Use `set.update` to include other `comms` (#4268)
Instead of coercing other collections to `set`s before performing a
union with the `set` of `comms`, which can take a bit longer, simply
call `comms.update(...)` on the provided collection. This avoids the
need to build a separate `set` before performing the union and still has
the same effect of including new values in `comms` where relevant.

Included a link to relevant Python docs for this method (for those
unfamiliar). Note that both `frozenset` and `set` are documented
together hence the reference to `frozenset` in the link even though it
applies to `set` also.

https://docs.python.org/3/library/stdtypes.html#frozenset.update