Add aggregated topic logs and log_event method (#4230)
This PR adds the ability to log events under a specified topic to a centralized ledger of events that lives on the scheduler. Currently the centralized ledger contains some administrative events which were previously logged using `Scheduler.log_event`. With the changes in this PR, the client and workers now also have a `log_event` method to add their own events to the ledger. Additionally, the client also has a `get_events` method which allows users to retrieve events from the ledger.

Closes https://github.com/dask/distributed/issues/4222