Precompute `hash` for `WorkerState`
Despite `self.address` being a simple `str`. It appears calling
`hash(...)` on it results in a bit of overhead. As we know that
`self.address` should not be changing, we can precompute this once in
`__init__` and simply `return` that result in `__hash__`. In all a
simple change that should cutdown on the time spent selecting a `Worker`
in `decide_worker`.