Only check for `c.client_key` once
Previously we looped through all `Client`s interested in a `TaskState`
and collected their `BatchedSend` comms object (if we had it). This
results in checking for the `Client` in our collection of `BatchedSend`
once and then retrieving it. IOW this is two calls to `__getitem__`. We
can cut this down to one call of `__getitem__` by simply trying to grab
the `Client` (whether it exists or not) and simply skipping over it if
we don't have it. The end result is the same, but the number of steps
performed is reduced.