Use a `list` collect `client_keys`
As `self.client_comms` is a `dict`, we know none of its keys will
repeat.  Similarly as we know `TaskState`'s `who_wants` is a set, none
of its values will repeat. So these can be added to a `list` without
creating duplicates. The only possible duplicate is the `client` key
passed into the function, which we can just check against the
`client_keys` once produced. So just do this instead thus avoiding a
bunch of checks for collisions that won't occur.