Only add `client` when it may not be included
In the case where all `Client`s are notified, there is no need to add
the `client` argument given to `report` as it is either in
`self.client_comms`, from which we included all keys, or it is not, in
which case we skip it. Thus we only need to try and added `client` when
only some of the `Client`s may have been notified, which may or may not
include the `client` argument given. Hence this branch is moved into the
`else` case.