Collect `report`'s `client_key`s in a `list` (#4275)
* Collect `comms` in `dict`

Instead of collecting a `set` of `comms` where `BatchedSend` objects
will need to be `hash`ed, use a `dict` with keys based on `Client` IDs
and their `BatchedSend` object as values.  This should still ensure that
we only pick up each `Client`'s `BatchedSend` object once. Though this
will hash the `Client` IDs (`str`s) instead of `BatchedSend` objects,
which should be more performant.

* Only check for `c.client_key` once

Previously we looped through all `Client`s interested in a `TaskState`
and collected their `BatchedSend` comms object (if we had it). This
results in checking for the `Client` in our collection of `BatchedSend`
once and then retrieving it. IOW this is two calls to `__getitem__`. We
can cut this down to one call of `__getitem__` by simply trying to grab
the `Client` (whether it exists or not) and simply skipping over it if
we don't have it. The end result is the same, but the number of steps
performed is reduced.

* Add blank line for readability

* Collect `Client` keys in a `set` instead

Per James' suggestion try collecting a `set` of `Client` keys instead.
This simplifies a bit of the logic earlier on. Plus we can fold the
handling of `Client`s whose comms we don't have into our existing
`for`-loop below.

* Use a `list` collect `client_keys`

As `self.client_comms` is a `dict`, we know none of its keys will
repeat.  Similarly as we know `TaskState`'s `who_wants` is a set, none
of its values will repeat. So these can be added to a `list` without
creating duplicates. The only possible duplicate is the `client` key
passed into the function, which we can just check against the
`client_keys` once produced. So just do this instead thus avoiding a
bunch of checks for collisions that won't occur.

* Add blank line for readability

* Assign `client_keys` directly

As `client_keys` is blank to begin with, just assign directly to it
instead of `extend`ing it.

* Only add `client` when it may not be included

In the case where all `Client`s are notified, there is no need to add
the `client` argument given to `report` as it is either in
`self.client_comms`, from which we included all keys, or it is not, in
which case we skip it. Thus we only need to try and added `client` when
only some of the `Client`s may have been notified, which may or may not
include the `client` argument given. Hence this branch is moved into the
`else` case.

* Only add keys that don't duplicate `client`

* Use `list` comprehensions instead of `for`-loops

* Flatten branch structure

* Run `black`

* Flip `client` cases