Make gather_dep robust to missing tasks (#4285)
Previously the gather_dep function assumed that tasks valid when
starting the function were still valid after an await call.  This may
not be the case.  This was fine, the rest of the work in this functions
was just about book-keeping for diagnostics, so errors didn't stop
things from halting.

Now we don't count the data sent for keys that we no longer know about.
This isn't entirely accurate, but should be fine.

Co-authored-by: James Bourbeau <jrbourbeau@users.noreply.github.com>