Drop typing of `self.idle`
As `self.idle` is actually a `SortedSet` instead of a `set`, we can't
type it. So revert the typing of `idle`. Though leave all other
`set`-based typing for other optimizations to be applied where possible.
Also leave the assignment of `self.idle` to `idle` as this generates the
attribute access code only once and we need this for both branches
anyways.