WIP - Optionally use offload executor in worker
Normally we have one separate thread pool executor for deserializing data
and another thread pool executor in a worker for execution.
Sometimes this presents a problem because some data types want to be
used in the thread in which they were created.  One example of this is
TensorFlow graphs, but there are others.

One way to resolve this is to reuse the same executor in both
situations, and ensure that it has only one thread.  This means that
execution and deserialization will block each other (not great) but that
user data will always be operated on in one thread only.

This commit implements that, and drafts up a small test.  However, it is
still broken because Dask can be clever in some situations and
deserialize directly on the event loop.  This happens for a few reasons
today:

1.  For small messages we deserialize on the event loop for performance
    reasons.  The user can control this with the
    distributed.comm.offload configuration value.

    I recommend the value of 1, meaning a single byte

2.  The scheduler-client comms intentionally do not offload today
    (grep for the `allow_offload=False`)