Annotate `WorkerState` for Cythonization (#4294)
Analogous to PR ( https://github.com/dask/distributed/pull/4290 ) except instead of `ClientState`, this is annotating `WorkerState` and usages thereof. This is a bit longer simply because of how many attributes `WorkerState` has. That said, this just uses the same strategy here as was used for `ClientState`.

* Coerce `bandwidth` to `float`

Typically `bandwidth` seems to be a `float` elsewhere in the code. So
coerce it to a `float` here as well to align with that practice.

* Use `ws` variable name for `WorkerState` objects

* Name `WorkerState` variable distinctly in closure

Avoids confusion with the variable name used outside this closure. Also
bypasses any potential leakage from the external scope.

* Assign selected `WorkerState` to variable

* Annotate `WorkerState` for Cythonization

* Annotate all `WorkerState` variables

* Prefix `WorkerState` attributes with `_`

Since these are effectively private given they are only available in
Cython/C, go ahead and mark them as such. We can then add Python
`@property`s on top of them to handle Python access to the original
attribute names with minimal disruption.

* Add Python-level `property`s for attributes

* Run `black`

* Add some `property.setter`s

Includes `setter`s for `occupancy` and `nbytes`.

* Create `list` from generator

This ensures Cython still uses `WorkerState` to annotate the variable
iterated over. Otherwise it constructs a generator with its own scope
where this is ignored.

* Relax `_name` to `object`

Apparently `_name` is sometimes an `int`. So relax typing around it to
just an `object`.