Some Cython fixes for `WorkerState` (#4321)
* Use `0` for `memory_limit` instead of `None`

We need an integer to pass to `WorkerState` so set this to `0` instead
of `None`.

* Use `list` comprehension to bind `WorkerState`

As Cython may not type things based on outer scope variables when using
a generator, convert the generator to a `list` comprehension, which does
respect the type of external variables used within the loop. This should
fix some attribute access errors seen here.