Use `-1` as `TaskState.nbytes` default
Ideally we would want to type `nbytes` as some kind of integer for
Cython optimization. However as we assign and work with `nbytes` being
`None` in some cases, this is not currently possible since a C integer
cannot be `None`. To fix that change the default value of `nbytes` to be
`-1`. This way it is still a valid integer and one we would never use
for `nbytes` unless it wasn't set. Thus it remains easy to check for and
won't overlap. Also it is something a C integer can be assigned.