Set `TaskPrefix.duration_average` to `-1`
Instead of using `None` for `TaskPrefix.duration_average`, set it `-1`.
This works better when typing `TaskPrefix.duration_average` as it can
always be floating point. This also works logically with this value as
it can't actually be negative unless it wasn't defined. Rework the logic
around this variable to ensure it is positive semi-definite.