Ensure `retire_workers` always `return`s a `dict`
As `workers_to_close` actually `return`s a `list` and that `list` could
be empty, we can wind up `return`ing an empty `list` instead of
`return`ing a `dict`. To fix that, we only return the `workers` object
when it is non-trivial (so will be replaced by a `dict`). In the trivial
case we just `return` an empty `dict`. This ensures we only `return` a
`dict` in this case, which fixes an issue that was cropping up when
compiling the scheduler with Cython.