Annotate `Task*` objects for Cythonization (#4302)
* Use `tsp` variable name for `TaskStreamPlugin`s

* Use `dts` for iterated `TaskState` variables

* Create `list` from generator

This ensures Cython still uses `TaskState` to annotate the variable
iterated over. Otherwise it constructs a generator with its own scope
where this is ignored.

* Use `-1` as `TaskState.nbytes` default

Ideally we would want to type `nbytes` as some kind of integer for
Cython optimization. However as we assign and work with `nbytes` being
`None` in some cases, this is not currently possible since a C integer
cannot be `None`. To fix that change the default value of `nbytes` to be
`-1`. This way it is still a valid integer and one we would never use
for `nbytes` unless it wasn't set. Thus it remains easy to check for and
won't overlap. Also it is something a C integer can be assigned.

* Assign `TaskState` instances to variables

* Annotate `TaskState` for Cythonization

* Annotate all `TaskState` variables

* Use closure to access `TaskState.priority`

This way we are able to inform Cython of the type and it is able to do
the work necessary to retrieve it. We will wind up reverting this later.

* Add `_` before all `TaskState` attributes

* Use `_` prefixed `TaskState` attributes throughout

* Add Python-level `property`s for attributes

* Add some `property.setter`s

Includes `setter`s for `processing_on` and `nbytes`.

* Drop recently added `TaskState.priority` closures

Now that we have properties accessible from `TaskState`, drop the
closures we added previously to access the typed values internally. This
should be equivalently performant and cut out a little bit of
boilerplate.

* Swap assignment order with `TaskPrefix`

Make sure to assign to the `TaskPrefix` variable, `tp`, first before
assigning to the `dict`. This should avoid the admittedly likely low
overhead of looking up the result in the dictionary when we already have
the value available.

* Assign `dask.config` query to a variable

Saves us need to fetch this twice. Also makes the code a bit more
readable. Finally may allow Cython optimizations on the variable later.

* Set `TaskPrefix.duration_average` to `-1`

Instead of using `None` for `TaskPrefix.duration_average`, set it `-1`.
This works better when typing `TaskPrefix.duration_average` as it can
always be floating point. This also works logically with this value as
it can't actually be negative unless it wasn't defined. Rework the logic
around this variable to ensure it is positive semi-definite.

* Annotate `TaskPrefix` for Cythonization

* Annotate `TaskPrefix` constructor as well

* Annotate all `TaskPrefix` variables

* Add `_` before all `TaskPrefix` attributes

* Add Python-level `property`s for attributes

* Use `Py_ssize_t` in `set_nbytes`

This allows Cython to perform C-level optimizations on these variables
and usages thereof.

* Use `tg` for `TaskGroup` variables

* Swap assignment order with `TaskGroup`

Make sure to assign to the `TaskGroup` variable, `tg`, first before
assigning to the `dict`. This should avoid the admittedly likely low
overhead of looking up the result in the dictionary when we already have
the value available.

* Create `list` from generator

This ensures Cython still uses `TaskGroup` to annotate the variable
iterated over. Otherwise it constructs a generator with its own scope
where this is ignored.

* Annotate `TaskGroup` for Cythonization

* Annotate `TaskGroup` constructor as well

* Set `TaskGroup.prefix` to `None` initially

* Annotate all `TaskGroup` variables

* Add `_` before all `TaskGroup` attributes

* Add Python-level `property`s for attributes